[#43] Update texture test
