+++ example/light/CMakeLists.txt
