+++ example/light/main.cpp
