+++ example/light/shaders/phong_light.vert
