+++ example/light/shaders/white_light_lamp.frag
