+++ neutrino/graphics/mesh.hpp
@@ -19,30 +19,33 @@ namespace framework::graphics
 /// @brief Mesh.
 ///
 /// Meshes contain vertex data (positions, normals, texture coordinates etc.)
-/// and triangles vertex indices.
+/// and vertex indices, which forms some kind of primitive to render.
 /// All vertex data must be in arrays of the same size.
-/// For example, if you have a mesh of 100 vertices then positions, normals and
-/// other arrays must be 100 in size.
 /// Data for i-th vertex is at index "i" in each array.
 ///
-/// For every vertex there can be a position, normal, tangent, color
-/// and up to 8 texture coordinates.
+/// For every vertex there can be a position, normal, tangent, color and up to 8 texture coordinates.
 ///
-/// The triangles are vertex indices of vertices the triangle made from.
-/// First three indices form a triangle, next three the other one and so on.
-/// So the indices array size must be divided by three.
-class Mesh
+/// Each mesh consist of group of vertex indices, called Submesh.
+/// At least one submesh must be defined to render something.
+class Mesh final
 {
 public:
+    /// @brief Represents what kind of primitive a submesh is describe.
     enum class PrimitiveType
     {
-        points,
-        lines,
-        line_strip,
-        line_loop,
-        triangles,
-        triangle_strip,
-        triangle_fan,
+        points,         ///< Each index is individual point.
+        lines,          ///< Vertices 0 and 1 are considered a line. Vertices 2 and 3 are considered a line. And so on.
+        line_strip,     ///< The adjacent vertices are considered lines. If you pass n vertices, you will get n-1 lines.
+        line_loop,      ///< As line strips, except that the first and last vertices are also used as a line.
+        triangles,      ///< Vertices 0, 1, and 2 form a triangle. Vertices 3, 4, and 5 form a triangle. And so on.
+        triangle_strip, ///<  Every group of 3 adjacent vertices forms a triangle. The face direction of the strip is
+                        ///<  determined by the winding of the first triangle. Each successive triangle will have its
+                        ///<  effective face order reversed, so the system compensates for that by testing it in the
+                        ///<  opposite way. A vertex stream of n length will generate n-2 triangles.
+
+        triangle_fan, ///< The first vertex is always held fixed. From there on, every group of 2 adjacent vertices form
+                      ///< a triangle with the first. So with a vertex stream, you get a list of triangles like so:
+                      ///< (0, 1, 2) (0, 2, 3), (0, 3, 4), etc. A vertex stream of n length will generate n-2 triangles.
     };
 
     using VertexData             = std::vector<math::Vector3f>;
@@ -68,7 +71,6 @@ public:
     ~Mesh();
 
     Mesh& operator=(const Mesh& other);
-
     Mesh& operator=(Mesh&& other) noexcept;
 
     /// @brief Assign new vertex positions to Mesh.
