+++ test/graphics/CMakeLists.txt
@@ -2,10 +2,10 @@ set(TESTS
     font
     image_bmp
     image_png
-    light
     mesh
     shader
     texture
+    #renderer
 )
 
 foreach(TEST ${TESTS})
