+++ test/graphics/renderer/CMakeLists.txt
@@ -15,3 +15,6 @@ add_test(NAME "${GROUP_NAME}/${TEST_NAME}" COMMAND ${TEST_NAME})
 set_tests_properties("${GROUP_NAME}/${TEST_NAME}" PROPERTIES LABELS ${GROUP_NAME})
 
 add_dependencies(${TESTS_TARGET} ${TEST_NAME})
+
+file(GLOB IMAGES "${TEST_DIR}/texture_img/*.png")
+file(COPY ${IMAGES} DESTINATION ${CMAKE_BINARY_DIR}/test/texture_img)
