+++ test/graphics/renderer/texture_img/lena_1024.png
