+++ test/graphics/renderer/texture_img/lena_32.png
